refactor(lsp): split server.ts into standalone analysis modules
server.ts had grown to ~460 lines mixing connection wiring with the
actual static analysis (tokenizing, call-site diagnostics, completion,
hover helpers), and it never leaves the file testable in isolation
(createConnection/.listen() run as import-time side effects). Pull the
pure logic into its own modules so it can be unit-tested directly:

- tokenize.ts — tokenizer, tree parser, call-site collector
- call-diagnostics.ts — required-:arg / arity diagnostics, with doc
  resolution injected as a callback so callers can supply a live
  session or a canned map
- symbols.ts — enclosingCallHead / symbolAt / markdownFor
- doc-args.ts — DocArg[] -> CompletionItem[] rendering
- load-mcp.ts — completions for load-mcp's bare toolkit-name string
  argument only; its :key plist form now falls through to the generic
  keywordCompletions path (see the paired interpreter commit)

server.ts itself shrinks to just the connection/session wiring.
